refactor(torghut): split profitability frontier payload ownership
