+++ services/torghut/scripts/search_consistent_profitability_frontier.py
@@ -133,9 +133,56 @@ from scripts.consistent_profitability_frontier.stress_metrics import (
     _second_oos_summary as _second_oos_summary,
     _holdout_oos_passed as _holdout_oos_passed,
 )
+from scripts.consistent_profitability_frontier.frontier_payload import (
+    _SAFE_EXACT_REPLAY_CANDIDATE_CAP as _SAFE_EXACT_REPLAY_CANDIDATE_CAP,
+    _build_economic_shortlist as _build_economic_shortlist,
+    _build_frontier_payload as _build_frontier_payload,
+    _build_frontier_workflow_states as _build_frontier_workflow_states,
+    _frontier_state_item as _frontier_state_item,
+    _rank_scored_candidates as _rank_scored_candidates,
+)
+from scripts.consistent_profitability_frontier.paper_probation import (
+    _PAPER_PROBATION_ACTIVITY_REPAIR_REASONS as _PAPER_PROBATION_ACTIVITY_REPAIR_REASONS,
+    _PAPER_PROBATION_CAPITAL_REPAIR_REASONS as _PAPER_PROBATION_CAPITAL_REPAIR_REASONS,
+    _PAPER_PROBATION_LIVE_PAPER_EVIDENCE_REQUIREMENTS as _PAPER_PROBATION_LIVE_PAPER_EVIDENCE_REQUIREMENTS,
+    _PAPER_PROBATION_LOSS_REPAIR_REASONS as _PAPER_PROBATION_LOSS_REPAIR_REASONS,
+    _PAPER_PROBATION_QUEUE_SURVIVAL_REASONS as _PAPER_PROBATION_QUEUE_SURVIVAL_REASONS,
+    _PAPER_PROBATION_SAFE_EVIDENCE_COLLECTION_PATH as _PAPER_PROBATION_SAFE_EVIDENCE_COLLECTION_PATH,
+    _PAPER_PROBATION_TAIL_RISK_REASONS as _PAPER_PROBATION_TAIL_RISK_REASONS,
+    _PAPER_PROBATION_TARGET_SCALE_QUANTUM as _PAPER_PROBATION_TARGET_SCALE_QUANTUM,
+    _bounded_sim_handoff_metadata as _bounded_sim_handoff_metadata,
+    _build_paper_probation_shortlist as _build_paper_probation_shortlist,
+    _candidate_artifact_refs as _candidate_artifact_refs,
+    _candidate_exact_replay_ledger_artifact_refs as _candidate_exact_replay_ledger_artifact_refs,
+    _candidate_exact_replay_parity_ok as _candidate_exact_replay_parity_ok,
+    _candidate_handoff_diagnostics as _candidate_handoff_diagnostics,
+    _candidate_metric_decimal as _candidate_metric_decimal,
+    _candidate_metric_value as _candidate_metric_value,
+    _candidate_post_cost_proof_blockers as _candidate_post_cost_proof_blockers,
+    _candidate_replay_tape_metadata_blockers as _candidate_replay_tape_metadata_blockers,
+    _candidate_runtime_ledger_count as _candidate_runtime_ledger_count,
+    _candidate_source_lineage_ok as _candidate_source_lineage_ok,
+    _paper_probation_notional_scale as _paper_probation_notional_scale,
+    _paper_probation_notional_scale_decimal as _paper_probation_notional_scale_decimal,
+    _paper_probation_repair_actions as _paper_probation_repair_actions,
+    _paper_probation_repair_plan as _paper_probation_repair_plan,
+    _paper_probation_required_actions as _paper_probation_required_actions,
+    _paper_probation_target_notional_scale as _paper_probation_target_notional_scale,
+    _paper_probation_target_progress as _paper_probation_target_progress,
+    _safe_decimal as _safe_decimal,
+)
+from scripts.consistent_profitability_frontier.repair_math import (
+    _LOSS_REPAIR_CAPITAL_SAFETY_BUFFER as _LOSS_REPAIR_CAPITAL_SAFETY_BUFFER,
+    _LOSS_REPAIR_DEFAULT_EXPOSURE_SCALE as _LOSS_REPAIR_DEFAULT_EXPOSURE_SCALE,
+    _LOSS_REPAIR_MIN_SCALE_QUANTUM as _LOSS_REPAIR_MIN_SCALE_QUANTUM,
+    _capital_repair_exposure_scale as _capital_repair_exposure_scale,
+    _decimal_or_none as _decimal_or_none,
+    _decimal_payload as _decimal_payload,
+    _reduced_exposure as _reduced_exposure,
+    _tightened_bps as _tightened_bps,
+)
 
 _LOCAL_ONLY_OVERRIDE_KEYS = frozenset({"normalization_regime"})
-_SAFE_EXACT_REPLAY_CANDIDATE_CAP = 6
 _SAFE_EXACT_REPLAY_EXPLOITATION_SLOTS = 4
 _SAFE_EXACT_REPLAY_EXPLORATION_SLOTS = 2
 _SAFE_LOCAL_EXACT_REPLAY_WORKERS = 2
@@ -176,9 +223,6 @@ _LOSS_REPAIR_STRATEGY_EXPOSURE_KEYS = (
     "max_notional_per_trade",
     "max_position_pct_equity",
 )
-_LOSS_REPAIR_DEFAULT_EXPOSURE_SCALE = Decimal("0.75")
-_LOSS_REPAIR_CAPITAL_SAFETY_BUFFER = Decimal("0.95")
-_LOSS_REPAIR_MIN_SCALE_QUANTUM = Decimal("0.000001")
 _CONSISTENCY_REPAIR_TRIGGER_REASONS = frozenset(
     {
         "active_day_ratio_below_min",
@@ -193,73 +237,6 @@ _CONSISTENCY_REPAIR_UNSAFE_REASONS = frozenset(
         "min_cash_below_min",
         "negative_cash_observation_count_above_max",
     }
-)
-_PAPER_PROBATION_CAPITAL_REPAIR_REASONS = frozenset(
-    {
-        "gross_exposure_pct_equity_above_max",
-        "min_cash_below_min",
-        "negative_cash_observation_count_above_max",
-    }
-)
-_PAPER_PROBATION_LOSS_REPAIR_REASONS = frozenset(
-    {
-        "daily_net_below_min",
-        "max_drawdown_above_max",
-        "worst_day_loss_above_max",
-    }
-)
-_PAPER_PROBATION_ACTIVITY_REPAIR_REASONS = frozenset(
-    {
-        "active_day_ratio_below_min",
-        "avg_daily_notional_below_min",
-        "best_day_share_above_max",
-        "profit_factor_below_min",
-        "second_oos_net_per_day_below_target",
-    }
-)
-_PAPER_PROBATION_TAIL_RISK_REASONS = frozenset(
-    {
-        "conformal_tail_risk_below_target",
-        "fill_survival_sample_count_below_min",
-        "fill_survival_rate_below_min",
-    }
-)
-_PAPER_PROBATION_QUEUE_SURVIVAL_REASONS = frozenset(
-    {
-        "fill_survival_sample_count_below_min",
-        "fill_survival_rate_below_min",
-        "required_min_queue_position_survival_sample_count",
-        "queue_position_survival_fill_curve_evidence_present_failed",
-        "queue_position_survival_sample_count_failed",
-        "queue_position_survival_fill_rate_failed",
-        "queue_position_survival_nonfill_opportunity_cost_present_failed",
-        "queue_position_survival_nonfill_opportunity_cost_bps_failed",
-        "queue_position_survival_fill_curve_evidence_missing",
-        "queue_position_survival_sample_count_zero",
-        "queue_position_survival_fill_rate_non_positive",
-        "queue_position_survival_queue_ahead_depletion_evidence_missing",
-        "queue_position_survival_queue_ahead_depletion_sample_count_zero",
-        "queue_position_survival_adjusted_fillable_ratio_non_positive",
-        "queue_position_survival_stress_net_pnl_non_positive",
-    }
-)
-_PAPER_PROBATION_TARGET_SCALE_QUANTUM = Decimal("0.000001")
-_PAPER_PROBATION_LIVE_PAPER_EVIDENCE_REQUIREMENTS = (
-    "real_runtime_trade_decisions",
-    "broker_order_submissions_and_status_events",
-    "broker_fill_events",
-    "post_cost_costs_tca_and_execution_shortfall",
-    "source_backed_runtime_ledger_lineage",
-    "closed_flat_position_proof",
-    "broker_runtime_ledger_reconciliation",
-)
-_PAPER_PROBATION_SAFE_EVIDENCE_COLLECTION_PATH = (
-    "preserve_final_promotion_gates_fail_closed",
-    "import_exact_replay_runtime_window_metadata_without_live_submit",
-    "run_bounded_live_paper_probe_with_configured_paper_account_only",
-    "materialize_source_backed_runtime_ledger_lineage_for_decisions_orders_fills_costs",
-    "verify_closed_flat_positions_and_broker_runtime_ledger_reconciliation",
-    "reevaluate_post_cost_runtime_profitability_before_any_final_promotion",
 )
 _CONSISTENCY_REPAIR_ENTRY_KEYS = (
     "max_entries_per_session",
@@ -2300,99 +2277,6 @@ def _strategy_item_from_configmap(
     return {}, {}
 
 
-def _decimal_or_none(value: Any) -> Decimal | None:
-    if value is None:
-        return None
-    text = str(value).strip()
-    if not text:
-        return None
-    try:
-        value = Decimal(text)
-    except (InvalidOperation, ValueError):
-        return None
-    if not value.is_finite():
-        return None
-    return value
-
-
-def _decimal_payload(value: Decimal) -> str:
-    text = format(value.normalize(), "f")
-    if "." in text:
-        text = text.rstrip("0").rstrip(".")
-    return text or "0"
-
-
-def _tightened_bps(value: Any, *, floor: Decimal) -> str | None:
-    current = _decimal_or_none(value)
-    if current is None or current <= 0:
-        return None
-    tightened = max(floor, (current * Decimal("0.75")).quantize(Decimal("0.01")))
-    if tightened >= current:
-        return None
-    return _decimal_payload(tightened)
-
-
-def _reduced_exposure(
-    value: Any,
-    *,
-    scale: Decimal = _LOSS_REPAIR_DEFAULT_EXPOSURE_SCALE,
-) -> str | None:
-    current = _decimal_or_none(value)
-    if current is None or current <= 0:
-        return None
-    if scale <= 0 or scale >= 1:
-        return None
-    reduced = (current * scale).quantize(
-        _LOSS_REPAIR_MIN_SCALE_QUANTUM, rounding=ROUND_DOWN
-    )
-    if reduced <= 0 or reduced >= current:
-        return None
-    return _decimal_payload(reduced)
-
-
-def _capital_repair_exposure_scale(
-    full_window_summary: Mapping[str, Any],
-    *,
-    policy_required_max_gross_exposure_pct_equity: Decimal | None = None,
-    policy_required_min_cash: Decimal | None = None,
-) -> Decimal:
-    scale = _LOSS_REPAIR_DEFAULT_EXPOSURE_SCALE
-    max_gross = _decimal_or_none(
-        full_window_summary.get("max_gross_exposure_pct_equity")
-    )
-    required_max_gross = _decimal_or_none(
-        full_window_summary.get("max_gross_exposure_pct_equity_required")
-        or full_window_summary.get("required_max_gross_exposure_pct_equity")
-    )
-    if required_max_gross is None or required_max_gross <= 0:
-        required_max_gross = Decimal("1")
-    if (
-        policy_required_max_gross_exposure_pct_equity is not None
-        and policy_required_max_gross_exposure_pct_equity > 0
-    ):
-        required_max_gross = min(
-            required_max_gross, policy_required_max_gross_exposure_pct_equity
-        )
-    if max_gross is not None and max_gross > required_max_gross:
-        gross_scale = (
-            required_max_gross / max_gross * _LOSS_REPAIR_CAPITAL_SAFETY_BUFFER
-        ).quantize(_LOSS_REPAIR_MIN_SCALE_QUANTUM, rounding=ROUND_DOWN)
-        scale = min(scale, max(_LOSS_REPAIR_MIN_SCALE_QUANTUM, gross_scale))
-
-    min_cash = _decimal_or_none(full_window_summary.get("min_cash"))
-    required_min_cash = _decimal_or_none(
-        full_window_summary.get("min_cash_required")
-        or full_window_summary.get("required_min_cash")
-    )
-    if required_min_cash is None:
-        required_min_cash = Decimal("0")
-    if policy_required_min_cash is not None:
-        required_min_cash = max(required_min_cash, policy_required_min_cash)
-    if min_cash is not None and min_cash < required_min_cash:
-        scale = min(scale, Decimal("0.5"))
-    return max(_LOSS_REPAIR_MIN_SCALE_QUANTUM, scale)
-
-
 def _loss_repair_trigger_reason(
     *,
     hard_vetoes: Sequence[Any],
@@ -2745,1356 +2629,6 @@ def _generate_consistency_repair_children(
         ),
     )
     return children
-
-
-def _rank_scored_candidates(scored: list[dict[str, Any]]) -> list[dict[str, Any]]:
-    if not scored:
-        return []
-
-    def promotion_grade_lower_bound(scorecard: Mapping[str, Any]) -> Decimal:
-        lower_bound = deployable_lower_bound_net_pnl_per_day(scorecard)
-        if lower_bound is None:
-            return Decimal("0")
-        if (
-            deployable_lower_bound_missing_count(scorecard) > 0
-            or deployable_proof_failed_gate_count(scorecard) > 0
-        ):
-            return Decimal("0")
-        return lower_bound
-
-    scorecards = {
-        str(item["candidate_id"]): build_scorecard(
-            candidate_id=str(item["candidate_id"]),
-            trading_day_count=int(item["full_window"]["trading_day_count"]),
-            net_pnl_per_day=Decimal(
-                str(item["objective_scorecard"]["net_pnl_per_day"])
-            ),
-            active_days=int(
-                Decimal(str(item["objective_scorecard"]["active_day_ratio"]))
-                * Decimal(str(item["full_window"]["trading_day_count"]))
-            ),
-            positive_days=int(
-                Decimal(str(item["objective_scorecard"]["positive_day_ratio"]))
-                * Decimal(str(item["full_window"]["trading_day_count"]))
-            ),
-            avg_filled_notional_per_day=Decimal(
-                str(item["objective_scorecard"]["avg_filled_notional_per_day"])
-            ),
-            avg_filled_notional_per_active_day=Decimal(
-                str(item["objective_scorecard"]["avg_filled_notional_per_active_day"])
-            ),
-            worst_day_loss=Decimal(str(item["objective_scorecard"]["worst_day_loss"])),
-            max_drawdown=Decimal(str(item["objective_scorecard"]["max_drawdown"])),
-            best_day_share=Decimal(str(item["objective_scorecard"]["best_day_share"])),
-            negative_day_count=int(item["objective_scorecard"]["negative_day_count"]),
-            rolling_3d_lower_bound=Decimal(
-                str(item["objective_scorecard"]["rolling_3d_lower_bound"])
-            ),
-            rolling_5d_lower_bound=Decimal(
-                str(item["objective_scorecard"]["rolling_5d_lower_bound"])
-            ),
-            regime_slice_pass_rate=Decimal(
-                str(item["objective_scorecard"]["regime_slice_pass_rate"])
-            ),
-            symbol_concentration_share=Decimal(
-                str(item["objective_scorecard"]["symbol_concentration_share"])
-            ),
-            entry_family_contribution_share=Decimal(
-                str(item["objective_scorecard"]["entry_family_contribution_share"])
-            ),
-            max_gross_exposure_pct_equity=Decimal(
-                str(
-                    item["objective_scorecard"].get(
-                        "max_gross_exposure_pct_equity", "0"
-                    )
-                )
-            ),
-            min_cash=Decimal(str(item["objective_scorecard"].get("min_cash", "0"))),
-            negative_cash_observation_count=int(
-                item["objective_scorecard"].get("negative_cash_observation_count") or 0
-            ),
-            fill_survival_sample_count=max(
-                _nonnegative_int_metric(
-                    item["objective_scorecard"].get(
-                        "delay_adjusted_depth_fill_survival_sample_count"
-                    )
-                ),
-                _nonnegative_int_metric(
-                    item["objective_scorecard"].get("fill_survival_sample_count")
-                ),
-                _nonnegative_int_metric(
-                    item["objective_scorecard"].get(
-                        "queue_position_survival_sample_count"
-                    )
-                ),
-            ),
-            fill_survival_rate=(
-                _optional_decimal(
-                    item["objective_scorecard"].get(
-                        "delay_adjusted_depth_fill_survival_rate"
-                    )
-                )
-                or _optional_decimal(
-                    item["objective_scorecard"].get("fill_survival_fill_rate")
-                )
-                or _optional_decimal(
-                    item["objective_scorecard"].get("queue_position_survival_fill_rate")
-                )
-                or Decimal("0")
-            ),
-            deployable_lower_bound_net_pnl_per_day=promotion_grade_lower_bound(
-                cast(Mapping[str, Any], item["objective_scorecard"])
-            ),
-        )
-        for item in scored
-    }
-    ranked_scorecards = rank_scorecards(
-        scorecards.values(),
-        veto_lookup={
-            str(item["candidate_id"]): tuple(
-                cast(list[str], item.get("hard_vetoes") or [])
-            )
-            for item in scored
-        },
-    )
-    ranked_lookup = {item.candidate_id: item for item in ranked_scorecards}
-    ranked_items = [dict(item) for item in scored]
-    for item in ranked_items:
-        ranked = ranked_lookup[str(item["candidate_id"])]
-        original_scorecard = dict(cast(Mapping[str, Any], item["objective_scorecard"]))
-        item["objective_scorecard"] = {
-            **original_scorecard,
-            **ranked.to_payload(),
-            "deployable_lower_bound_missing_count": deployable_lower_bound_missing_count(
-                original_scorecard
-            ),
-            "deployable_lower_bound_failed_gate_count": (
-                deployable_proof_failed_gate_count(original_scorecard)
-            ),
-        }
-        item["ranking"] = {
-            "method": "pareto_frontier_v2",
-            "pareto_tier": ranked.pareto_tier,
-            "tie_breaker_score": str(ranked.tie_breaker_score),
-            "vetoed": bool(ranked.veto_reasons),
-        }
-    ranked_items.sort(
-        key=lambda item: (
-            bool(item["ranking"]["vetoed"]),
-            int(item["ranking"]["pareto_tier"]),
-            -Decimal(str(item["ranking"]["tie_breaker_score"])),
-            -Decimal(str(item["full_window"]["net_per_day"])),
-        )
-    )
-    return ranked_items
-
-
-def _safe_decimal(value: Any) -> Decimal:
-    if value in (None, ""):
-        return Decimal("0")
-    try:
-        return Decimal(str(value))
-    except (InvalidOperation, ValueError):
-        return Decimal("0")
-
-
-def _candidate_metric_value(
-    item: Mapping[str, Any],
-    *,
-    scorecard_key: str,
-    full_window_key: str,
-    default: str = "0",
-) -> Any:
-    scorecard = _mapping(item.get("objective_scorecard"))
-    raw_value = scorecard.get(scorecard_key)
-    if raw_value not in (None, ""):
-        return raw_value
-    full_window = _mapping(item.get("full_window"))
-    raw_value = full_window.get(full_window_key)
-    if raw_value not in (None, ""):
-        return raw_value
-    return default
-
-
-def _build_economic_shortlist(
-    ranked_items: Sequence[Mapping[str, Any]], *, top_n: int
-) -> list[dict[str, Any]]:
-    visible_items = [
-        item
-        for item in ranked_items
-        if _safe_decimal(
-            _candidate_metric_value(
-                item,
-                scorecard_key="net_pnl_per_day",
-                full_window_key="net_per_day",
-            )
-        )
-        != Decimal("0")
-        or item.get("exact_replay_ledger_artifact_ref")
-        or item.get("replay_artifact_refs")
-    ]
-    visible_items.sort(
-        key=lambda item: (
-            -_safe_decimal(
-                _candidate_metric_value(
-                    item,
-                    scorecard_key="net_pnl_per_day",
-                    full_window_key="net_per_day",
-                )
-            ),
-            -_safe_decimal(
-                _candidate_metric_value(
-                    item,
-                    scorecard_key="net_pnl",
-                    full_window_key="net_pnl",
-                )
-            ),
-            str(item.get("candidate_id") or ""),
-        )
-    )
-    shortlist: list[dict[str, Any]] = []
-    for item in visible_items[: max(1, int(top_n))]:
-        ranking = _mapping(item.get("ranking"))
-        hard_vetoes = [
-            str(reason) for reason in cast(Sequence[Any], item.get("hard_vetoes") or ())
-        ]
-        shortlist.append(
-            {
-                "candidate_id": str(item.get("candidate_id") or ""),
-                "strategy_name": str(item.get("strategy_name") or ""),
-                "family": str(item.get("family") or ""),
-                "net_pnl_per_day": str(
-                    _candidate_metric_value(
-                        item,
-                        scorecard_key="net_pnl_per_day",
-                        full_window_key="net_per_day",
-                    )
-                ),
-                "net_pnl": str(
-                    _candidate_metric_value(
-                        item,
-                        scorecard_key="net_pnl",
-                        full_window_key="net_pnl",
-                    )
-                ),
-                "active_day_ratio": str(
-                    _candidate_metric_value(
-                        item,
-                        scorecard_key="active_day_ratio",
-                        full_window_key="active_ratio",
-                    )
-                ),
-                "positive_day_ratio": str(
-                    _candidate_metric_value(
-                        item,
-                        scorecard_key="positive_day_ratio",
-                        full_window_key="positive_day_ratio",
-                    )
-                ),
-                "max_drawdown": str(
-                    _candidate_metric_value(
-                        item,
-                        scorecard_key="max_drawdown",
-                        full_window_key="max_drawdown",
-                    )
-                ),
-                "worst_day_loss": str(
-                    _candidate_metric_value(
-                        item,
-                        scorecard_key="worst_day_loss",
-                        full_window_key="worst_day_loss",
-                    )
-                ),
-                "max_gross_exposure_pct_equity": str(
-                    _candidate_metric_value(
-                        item,
-                        scorecard_key="max_gross_exposure_pct_equity",
-                        full_window_key="max_gross_exposure_pct_equity",
-                    )
-                ),
-                "min_cash": str(
-                    _candidate_metric_value(
-                        item,
-                        scorecard_key="min_cash",
-                        full_window_key="min_cash",
-                    )
-                ),
-                "exact_replay_ledger_artifact_ref": str(
-                    item.get("exact_replay_ledger_artifact_ref") or ""
-                ),
-                "replay_artifact_refs": [
-                    str(ref)
-                    for ref in cast(
-                        Sequence[Any], item.get("replay_artifact_refs") or ()
-                    )
-                ],
-                "hard_vetoes": hard_vetoes,
-                "vetoed": bool(ranking.get("vetoed") or hard_vetoes),
-                "pareto_tier": ranking.get("pareto_tier"),
-            }
-        )
-    return shortlist
-
-
-def _candidate_artifact_refs(item: Mapping[str, Any]) -> list[str]:
-    refs: list[str] = []
-    exact_replay_ref = str(item.get("exact_replay_ledger_artifact_ref") or "").strip()
-    if exact_replay_ref:
-        refs.append(exact_replay_ref)
-    refs.extend(
-        str(ref).strip()
-        for ref in cast(Sequence[Any], item.get("replay_artifact_refs") or ())
-        if str(ref).strip()
-    )
-    return list(dict.fromkeys(refs))
-
-
-def _candidate_exact_replay_ledger_artifact_refs(item: Mapping[str, Any]) -> list[str]:
-    refs: list[str] = []
-    scorecard = _mapping(item.get("objective_scorecard"))
-    for source in (item, scorecard):
-        for key in ("exact_replay_ledger_artifact_ref",):
-            ref = str(source.get(key) or "").strip()
-            if ref:
-                refs.append(ref)
-        for key in ("exact_replay_ledger_artifact_refs",):
-            raw_refs = source.get(key)
-            if isinstance(raw_refs, str):
-                ref = raw_refs.strip()
-                if ref:
-                    refs.append(ref)
-                continue
-            for raw_ref in cast(Sequence[Any], raw_refs or ()):
-                ref = str(raw_ref).strip()
-                if ref:
-                    refs.append(ref)
-    for raw_ref in cast(Sequence[Any], item.get("replay_artifact_refs") or ()):
-        ref = str(raw_ref).strip()
-        ref_lower = ref.lower()
-        if ref and (
-            "exact-replay-ledger" in ref_lower or "exact_replay_ledger" in ref_lower
-        ):
-            refs.append(ref)
-    return list(dict.fromkeys(refs))
-
-
-def _candidate_runtime_ledger_count(item: Mapping[str, Any], *keys: str) -> int:
-    scorecard = _mapping(item.get("objective_scorecard"))
-    values: list[int] = []
-    for source in (item, scorecard):
-        values.extend(_nonnegative_int_metric(source.get(key)) for key in keys)
-    return max(values, default=0)
-
-
-def _paper_probation_required_actions(hard_vetoes: Sequence[str]) -> list[str]:
-    reasons = {str(reason) for reason in hard_vetoes}
-    actions = ["collect_runtime_ledger_paper_evidence"]
-    if reasons & _PAPER_PROBATION_CAPITAL_REPAIR_REASONS:
-        actions.append("apply_capital_repair_sizing_before_paper_orders")
-    if reasons & _PAPER_PROBATION_LOSS_REPAIR_REASONS:
-        actions.append("tighten_loss_controls_before_paper_orders")
-    if reasons & _PAPER_PROBATION_ACTIVITY_REPAIR_REASONS:
-        actions.append("collect_consistency_and_activity_evidence")
-    if reasons & _PAPER_PROBATION_TAIL_RISK_REASONS:
-        actions.append("collect_tail_risk_and_fill_survival_evidence")
-    if reasons & _PAPER_PROBATION_QUEUE_SURVIVAL_REASONS:
-        actions.append("collect_queue_position_survival_fill_curve_evidence")
-    if "stale_dataset_snapshot" in reasons:
-        actions.append("refresh_dataset_snapshot_before_paper_orders")
-    return actions
-
-
-def _paper_probation_notional_scale_decimal(
-    *,
-    item: Mapping[str, Any],
-    objective_veto_policy: ObjectiveVetoPolicy,
-    hard_vetoes: Sequence[str],
-) -> Decimal:
-    if not (
-        {str(reason) for reason in hard_vetoes}
-        & _PAPER_PROBATION_CAPITAL_REPAIR_REASONS
-    ):
-        return Decimal("1")
-    return _capital_repair_exposure_scale(
-        {
-            "max_gross_exposure_pct_equity": _candidate_metric_value(
-                item,
-                scorecard_key="max_gross_exposure_pct_equity",
-                full_window_key="max_gross_exposure_pct_equity",
-            ),
-            "max_gross_exposure_pct_equity_required": str(
-                objective_veto_policy.required_max_gross_exposure_pct_equity
-            ),
-            "min_cash": _candidate_metric_value(
-                item,
-                scorecard_key="min_cash",
-                full_window_key="min_cash",
-            ),
-            "min_cash_required": str(objective_veto_policy.required_min_cash),
-        },
-        policy_required_max_gross_exposure_pct_equity=(
-            objective_veto_policy.required_max_gross_exposure_pct_equity
-        ),
-        policy_required_min_cash=objective_veto_policy.required_min_cash,
-    )
-
-
-def _paper_probation_notional_scale(
-    *,
-    item: Mapping[str, Any],
-    objective_veto_policy: ObjectiveVetoPolicy,
-    hard_vetoes: Sequence[str],
-) -> str:
-    scale = _paper_probation_notional_scale_decimal(
-        item=item,
-        objective_veto_policy=objective_veto_policy,
-        hard_vetoes=hard_vetoes,
-    )
-    return _decimal_payload(scale)
-
-
-def _paper_probation_target_notional_scale(
-    *,
-    capital_repaired_net_pnl_per_day: Decimal,
-    target_net_pnl_per_day: Decimal,
-) -> Decimal:
-    if target_net_pnl_per_day <= 0:
-        return Decimal("1")
-    if capital_repaired_net_pnl_per_day <= 0:
-        return Decimal("0")
-    if capital_repaired_net_pnl_per_day >= target_net_pnl_per_day:
-        return Decimal("1")
-    return (target_net_pnl_per_day / capital_repaired_net_pnl_per_day).quantize(
-        _PAPER_PROBATION_TARGET_SCALE_QUANTUM,
-        rounding=ROUND_CEILING,
-    )
-
-
-def _candidate_metric_decimal(
-    item: Mapping[str, Any],
-    *,
-    scorecard_key: str,
-    full_window_key: str,
-) -> Decimal:
-    return _safe_decimal(
-        _candidate_metric_value(
-            item,
-            scorecard_key=scorecard_key,
-            full_window_key=full_window_key,
-        )
-    )
-
-
-def _candidate_source_lineage_ok(item: Mapping[str, Any]) -> bool:
-    replay_lineage = _mapping(item.get("replay_lineage"))
-    candidate_key = _mapping(item.get("candidate_evaluation_key_payload"))
-    return bool(
-        str(item.get("dataset_snapshot_id") or "").strip()
-        and str(replay_lineage.get("lineage_hash") or "").strip()
-        and str(
-            item.get("candidate_evaluation_key")
-            or candidate_key.get("candidate_evaluation_key")
-            or ""
-        ).strip()
-    )
-
-
-def _candidate_replay_tape_metadata_blockers(item: Mapping[str, Any]) -> list[str]:
-    candidate_key = _mapping(item.get("candidate_evaluation_key_payload"))
-    replay_tape = _mapping(candidate_key.get("replay_tape"))
-    if not replay_tape:
-        replay_tape = _mapping(item.get("replay_tape"))
-    if not replay_tape:
-        return []
-
-    required_fields = (
-        "content_sha256",
-        "dataset_snapshot_ref",
-        "source_query_digest",
-        "feature_schema_hash",
-        "cost_model_hash",
-        "strategy_family",
-        "replay_cache_key",
-    )
-    blockers = [
-        f"missing_replay_tape_{field}"
-        for field in required_fields
-        if not str(replay_tape.get(field) or "").strip()
-    ]
-    if int(replay_tape.get("selected_row_count") or 0) <= 0:
-        blockers.append("missing_replay_tape_selected_rows")
-    cache_identity = _mapping(replay_tape.get("cache_identity"))
-    blockers.extend(
-        str(blocker)
-        for blocker in cast(Sequence[Any], cache_identity.get("blockers") or ())
-        if str(blocker).strip()
-    )
-    return list(dict.fromkeys(blockers))
-
-
-def _candidate_post_cost_proof_blockers(item: Mapping[str, Any]) -> list[str]:
-    blockers: list[str] = []
-    pnl_basis = str(item.get("runtime_ledger_pnl_basis") or "").strip()
-    if pnl_basis != POST_COST_PNL_BASIS:
-        blockers.append("missing_post_cost_pnl_basis")
-
-    cost_basis_counts = item.get("runtime_ledger_cost_basis_counts")
-    cost_basis_count = _candidate_runtime_ledger_count(
-        item,
-        "runtime_ledger_cost_basis_count",
-        "exact_replay_ledger_cost_basis_count",
-    )
-    if isinstance(cost_basis_counts, Mapping):
-        parsed_cost_basis_count = 0
-        for value in cost_basis_counts.values():
-            try:
-                parsed_cost_basis_count += int(value or 0)
-            except (TypeError, ValueError):
-                continue
-        cost_basis_count = max(
-            cost_basis_count,
-            parsed_cost_basis_count,
-        )
-    if cost_basis_count <= 0 and not str(item.get("cost_basis") or "").strip():
-        blockers.append("missing_cost_basis")
-
-    expectancy_value = item.get("runtime_ledger_post_cost_expectancy_bps")
-    if expectancy_value in (None, ""):
-        blockers.append("missing_post_cost_expectancy_bps")
-    else:
-        expectancy = _safe_decimal(expectancy_value)
-        if expectancy <= Decimal("0"):
-            blockers.append("non_positive_post_cost_expectancy_bps")
-    return blockers
-
-
-def _candidate_exact_replay_parity_ok(
-    item: Mapping[str, Any],
-    *,
-    artifact_refs: Sequence[str],
-    exact_replay_ledger_row_count: int,
-    exact_replay_ledger_fill_count: int,
-) -> bool:
-    scorecard = _mapping(item.get("objective_scorecard"))
-    has_exact_ref = any(
-        "exact" in ref.lower() and "ledger" in ref.lower() for ref in artifact_refs
-    )
-    has_exact_ref = has_exact_ref or bool(
-        str(
-            item.get("exact_replay_ledger_artifact_ref")
-            or scorecard.get("exact_replay_ledger_artifact_ref")
-            or ""
-        ).strip()
-    )
-    return bool(
-        has_exact_ref
-        and exact_replay_ledger_row_count > 0
-        and exact_replay_ledger_fill_count > 0
-    )
-
-
-def _candidate_handoff_diagnostics(
-    item: Mapping[str, Any],
-    *,
-    hard_vetoes: Sequence[str],
-    artifact_refs: Sequence[str],
-    exact_replay_ledger_row_count: int,
-    exact_replay_ledger_fill_count: int,
-    source_lineage_ok: bool,
-    exact_replay_parity_ok: bool,
-) -> list[dict[str, Any]]:
-    reasons = {str(reason) for reason in hard_vetoes}
-    diagnostics: list[dict[str, Any]] = []
-
-    if reasons & {
-        "active_day_ratio_below_min",
-        "min_active_days_below_min",
-        "positive_day_ratio_below_min",
-        "second_oos_net_per_day_below_target",
-    } or _candidate_metric_decimal(
-        item,
-        scorecard_key="active_day_ratio",
-        full_window_key="active_ratio",
-    ) <= Decimal("0"):
-        diagnostics.append(
-            {
-                "category": "insufficient_days",
-                "reasons": sorted(
-                    reasons
-                    & {
-                        "active_day_ratio_below_min",
-                        "min_active_days_below_min",
-                        "positive_day_ratio_below_min",
-                        "second_oos_net_per_day_below_target",
-                    }
-                ),
-            }
-        )
-
-    cost_capacity_reasons = reasons & {
-        "avg_daily_notional_below_min",
-        "avg_filled_notional_per_active_day_below_min",
-        "gross_exposure_pct_equity_above_max",
-        "min_cash_below_min",
-        "negative_cash_observation_count_above_max",
-        "adv_capacity_below_min",
-        "capacity_below_target",
-        "fill_survival_rate_below_min",
-        "fill_survival_sample_count_below_min",
-        "profit_factor_below_min",
-    }
-    if cost_capacity_reasons:
-        diagnostics.append(
-            {
-                "category": "cost_adv_capacity_blockers",
-                "reasons": sorted(cost_capacity_reasons),
-            }
-        )
-
-    open_position_count = _candidate_runtime_ledger_count(
-        item,
-        "runtime_ledger_open_position_count",
-        "exact_replay_ledger_open_position_count",
-    )
-    if open_position_count > 0:
-        diagnostics.append(
-            {
-                "category": "open_replay_positions",
-                "open_position_count": open_position_count,
-            }
-        )
-
-    if "best_day_share_above_max" in reasons:
-        diagnostics.append(
-            {
-                "category": "best_day_concentration",
-                "reasons": ["best_day_share_above_max"],
-                "best_day_share_of_total_pnl": str(
-                    _candidate_metric_value(
-                        item,
-                        scorecard_key="best_day_share_of_total_pnl",
-                        full_window_key="best_day_share_of_total_pnl",
-                    )
-                ),
-            }
-        )
-
-    if not source_lineage_ok:
-        diagnostics.append(
-            {
-                "category": "missing_source_lineage",
-                "missing": [
-                    name
-                    for name, present in (
-                        (
-                            "dataset_snapshot_id",
-                            bool(str(item.get("dataset_snapshot_id") or "").strip()),
-                        ),
-                        (
-                            "replay_lineage.lineage_hash",
-                            bool(
-                                str(
-                                    _mapping(item.get("replay_lineage")).get(
-                                        "lineage_hash"
-                                    )
-                                    or ""
-                                ).strip()
-                            ),
-                        ),
-                        (
-                            "candidate_evaluation_key",
-                            bool(
-                                str(
-                                    item.get("candidate_evaluation_key")
-                                    or _mapping(
-                                        item.get("candidate_evaluation_key_payload")
-                                    ).get("candidate_evaluation_key")
-                                    or ""
-                                ).strip()
-                            ),
-                        ),
-                    )
-                    if not present
-                ],
-            }
-        )
-
-    if not exact_replay_parity_ok:
-        diagnostics.append(
-            {
-                "category": "failed_exact_replay_parity",
-                "artifact_refs": list(artifact_refs),
-                "exact_replay_ledger_artifact_row_count": (
-                    exact_replay_ledger_row_count
-                ),
-                "exact_replay_ledger_artifact_fill_count": (
-                    exact_replay_ledger_fill_count
-                ),
-            }
-        )
-
-    if (
-        artifact_refs
-        and exact_replay_ledger_row_count > 0
-        and exact_replay_ledger_fill_count > 0
-    ):
-        replay_tape_blockers = _candidate_replay_tape_metadata_blockers(item)
-        if replay_tape_blockers:
-            diagnostics.append(
-                {
-                    "category": "missing_replay_tape_metadata",
-                    "blockers": replay_tape_blockers,
-                }
-            )
-
-        post_cost_blockers = _candidate_post_cost_proof_blockers(item)
-        if post_cost_blockers:
-            diagnostics.append(
-                {
-                    "category": "missing_post_cost_basis",
-                    "blockers": post_cost_blockers,
-                }
-            )
-
-    return diagnostics
-
-
-def _bounded_sim_handoff_metadata(
-    *,
-    item: Mapping[str, Any],
-    paper_probation_allowed: bool,
-    source_lineage_ok: bool,
-    exact_replay_parity_ok: bool,
-    diagnostics: Sequence[Mapping[str, Any]],
-) -> dict[str, Any]:
-    evidence_collection_ok = bool(
-        paper_probation_allowed and source_lineage_ok and exact_replay_parity_ok
-    )
-    blockers = [
-        str(diagnostic.get("category") or "")
-        for diagnostic in diagnostics
-        if str(diagnostic.get("category") or "")
-    ]
-    if not blockers and not evidence_collection_ok:
-        blockers.append("source_or_replay_preconditions_not_met")
-    return {
-        "schema_version": "torghut.frontier-bounded-sim-handoff.v1",
-        "status": (
-            "evidence_collection_ready_promotion_blocked"
-            if evidence_collection_ok
-            else "handoff_blocked_until_preconditions_pass"
-        ),
-        "authority": "bounded_sim_evidence_collection_only",
-        "candidate_id": str(item.get("candidate_id") or ""),
-        "evidence_collection_ok": evidence_collection_ok,
-        "promotion_allowed": False,
-        "final_promotion_allowed": False,
-        "promotion_blockers": [
-            "source_backed_runtime_ledger_authority_required",
-            "live_paper_probation_runtime_evidence_required",
-            "unchanged_final_promotion_gates_required",
-        ],
-        "handoff_blockers": blockers,
-        "source_lineage_ok": source_lineage_ok,
-        "exact_replay_parity_ok": exact_replay_parity_ok,
-        "diagnostics": [dict(diagnostic) for diagnostic in diagnostics],
-    }
-
-
-def _paper_probation_repair_actions(
-    *,
-    blockers: Sequence[str],
-    hard_vetoes: Sequence[str],
-) -> list[str]:
-    actions = list(_paper_probation_required_actions(hard_vetoes))
-    blocker_set = {str(blocker) for blocker in blockers}
-    if blocker_set & {
-        "missing_exact_replay_ledger_artifact",
-        "missing_exact_replay_ledger_row_count",
-        "missing_exact_replay_ledger_fill_count",
-        "failed_exact_replay_parity",
-    }:
-        actions.append("produce_authoritative_exact_replay_ledger")
-    if "missing_source_lineage" in blocker_set:
-        actions.append("materialize_source_lineage_receipt")
-    if any(blocker.startswith("missing_replay_tape_") for blocker in blocker_set):
-        actions.append("refresh_replay_tape_metadata")
-    if blocker_set & {
-        "missing_post_cost_pnl_basis",
-        "missing_cost_basis",
-        "missing_post_cost_expectancy_bps",
-        "non_positive_post_cost_expectancy_bps",
-    }:
-        actions.append("rematerialize_post_cost_runtime_ledger")
-    if "open_replay_positions" in blocker_set:
-        actions.append("close_or_exclude_open_replay_position_windows")
-    if "proof_only_full_window_replay_not_probation_authority" in blocker_set:
-        actions.append("rerun_exact_replay_with_authoritative_runtime_events")
-    if blocker_set & _PAPER_PROBATION_QUEUE_SURVIVAL_REASONS:
-        actions.append("collect_queue_position_survival_fill_curve_evidence")
-    actions.append("keep_final_promotion_gates_fail_closed")
-    return list(dict.fromkeys(actions))
-
-
-def _paper_probation_target_progress(
-    *,
-    net_pnl_per_day: Decimal,
-    target_net_pnl_per_day: Decimal,
-) -> Decimal:
-    if target_net_pnl_per_day <= 0:
-        return Decimal("0")
-    return (net_pnl_per_day / target_net_pnl_per_day).quantize(Decimal("0.0001"))
-
-
-def _paper_probation_repair_plan(
-    *,
-    item: Mapping[str, Any],
-    blockers: Sequence[str],
-    hard_vetoes: Sequence[str],
-    handoff_diagnostics: Sequence[Mapping[str, Any]],
-    net_pnl_per_day: Decimal,
-    target_net_pnl_per_day: Decimal,
-    capital_repair_notional_scale: Decimal,
-    capital_repaired_net_pnl_per_day: Decimal,
-    target_notional_scale_after_capital_repair: Decimal,
-) -> dict[str, Any]:
-    repair_actions = _paper_probation_repair_actions(
-        blockers=blockers,
-        hard_vetoes=hard_vetoes,
-    )
-    target_shortfall = max(target_net_pnl_per_day - net_pnl_per_day, Decimal("0"))
-    capital_repaired_target_shortfall = max(
-        target_net_pnl_per_day - capital_repaired_net_pnl_per_day,
-        Decimal("0"),
-    )
-    target_scale_required = target_notional_scale_after_capital_repair > Decimal("1")
-    if target_scale_required:
-        repair_actions.append(
-            "validate_target_notional_scale_capacity_before_paper_orders"
-        )
-    repair_required = bool(blockers)
-    return {
-        "schema_version": "torghut.paper-probation-repair-plan.v1",
-        "status": "repair_required_before_paper_evidence_collection"
-        if repair_required
-        else "ready_for_bounded_paper_evidence_collection",
-        "candidate_id": str(item.get("candidate_id") or ""),
-        "target_net_pnl_per_day": _decimal_payload(target_net_pnl_per_day),
-        "observed_net_pnl_per_day": _decimal_payload(net_pnl_per_day),
-        "target_shortfall": _decimal_payload(target_shortfall),
-        "capital_repair_notional_scale": _decimal_payload(
-            capital_repair_notional_scale
-        ),
-        "capital_repaired_net_pnl_per_day": _decimal_payload(
-            capital_repaired_net_pnl_per_day
-        ),
-        "capital_repaired_target_shortfall": _decimal_payload(
-            capital_repaired_target_shortfall
-        ),
-        "target_notional_scale_after_capital_repair": _decimal_payload(
-            target_notional_scale_after_capital_repair
-        ),
-        "target_scale_required": target_scale_required,
-        "target_scale_authority": "planning_only_requires_bounded_paper_validation",
-        "live_paper_evidence_requirements": list(
-            _PAPER_PROBATION_LIVE_PAPER_EVIDENCE_REQUIREMENTS
-        ),
-        "safe_evidence_collection_path": list(
-            _PAPER_PROBATION_SAFE_EVIDENCE_COLLECTION_PATH
-        ),
-        "target_progress_ratio": _decimal_payload(
-            _paper_probation_target_progress(
-                net_pnl_per_day=net_pnl_per_day,
-                target_net_pnl_per_day=target_net_pnl_per_day,
-            )
-        ),
-        "repair_required": repair_required,
-        "repairable_for_evidence_collection": net_pnl_per_day > 0,
-        "repair_actions": repair_actions,
-        "repair_blockers": list(dict.fromkeys(str(blocker) for blocker in blockers)),
-        "diagnostics": [dict(diagnostic) for diagnostic in handoff_diagnostics],
-        "live_capital_authorized": False,
-        "promotion_allowed": False,
-        "final_promotion_authorized": False,
-        "final_promotion_allowed": False,
-        "authorization_scope": "evidence_collection_repair_only",
-    }
-
-
-def _build_paper_probation_shortlist(
-    ranked_items: Sequence[Mapping[str, Any]],
-    *,
-    top_n: int,
-    objective_veto_policy: ObjectiveVetoPolicy,
-    target_net_pnl_per_day: Decimal = Decimal("500"),
-) -> list[dict[str, Any]]:
-    visible_items = [
-        item
-        for item in ranked_items
-        if _safe_decimal(
-            _candidate_metric_value(
-                item,
-                scorecard_key="net_pnl_per_day",
-                full_window_key="net_per_day",
-            )
-        )
-        > Decimal("0")
-    ]
-
-    shortlist_candidates: list[
-        tuple[bool, bool, Decimal, Decimal, Decimal, Decimal, str, dict[str, Any]]
-    ] = []
-    for item in visible_items:
-        hard_vetoes = [
-            str(reason) for reason in cast(Sequence[Any], item.get("hard_vetoes") or ())
-        ]
-        artifact_refs = _candidate_exact_replay_ledger_artifact_refs(item)
-        exact_replay_ledger_row_count = _candidate_runtime_ledger_count(
-            item,
-            "exact_replay_ledger_artifact_row_count",
-        )
-        exact_replay_ledger_fill_count = _candidate_runtime_ledger_count(
-            item,
-            "exact_replay_ledger_artifact_fill_count",
-        )
-        blockers: list[str] = []
-        if not artifact_refs:
-            blockers.append("missing_exact_replay_ledger_artifact")
-        if exact_replay_ledger_row_count <= 0:
-            blockers.append("missing_exact_replay_ledger_row_count")
-        if exact_replay_ledger_fill_count <= 0:
-            blockers.append("missing_exact_replay_ledger_fill_count")
-        screening = _mapping(item.get("screening"))
-        staged_search = _mapping(item.get("staged_search"))
-        if (
-            bool(item.get("exact_replay_ledger_artifact_proof_only"))
-            or bool(item.get("runtime_ledger_artifact_proof_only"))
-            or bool(screening.get("proof_only_full_window_replay_captured"))
-            or str(staged_search.get("stage") or "").endswith("_full_window_proof")
-        ):
-            blockers.append("proof_only_full_window_replay_not_probation_authority")
-        source_lineage_ok = _candidate_source_lineage_ok(item)
-        exact_replay_parity_ok = _candidate_exact_replay_parity_ok(
-            item,
-            artifact_refs=artifact_refs,
-            exact_replay_ledger_row_count=exact_replay_ledger_row_count,
-            exact_replay_ledger_fill_count=exact_replay_ledger_fill_count,
-        )
-        if not source_lineage_ok:
-            blockers.append("missing_source_lineage")
-        if not exact_replay_parity_ok:
-            blockers.append("failed_exact_replay_parity")
-        if (
-            artifact_refs
-            and exact_replay_ledger_row_count > 0
-            and exact_replay_ledger_fill_count > 0
-        ):
-            blockers.extend(_candidate_replay_tape_metadata_blockers(item))
-            blockers.extend(_candidate_post_cost_proof_blockers(item))
-        open_position_count = _candidate_runtime_ledger_count(
-            item,
-            "runtime_ledger_open_position_count",
-            "exact_replay_ledger_open_position_count",
-        )
-        if open_position_count > 0:
-            blockers.append("open_replay_positions")
-        blockers = list(dict.fromkeys(blockers))
-        handoff_diagnostics = _candidate_handoff_diagnostics(
-            item,
-            hard_vetoes=hard_vetoes,
-            artifact_refs=artifact_refs,
-            exact_replay_ledger_row_count=exact_replay_ledger_row_count,
-            exact_replay_ledger_fill_count=exact_replay_ledger_fill_count,
-            source_lineage_ok=source_lineage_ok,
-            exact_replay_parity_ok=exact_replay_parity_ok,
-        )
-        paper_probation_allowed = not blockers
-        bounded_sim_handoff = _bounded_sim_handoff_metadata(
-            item=item,
-            paper_probation_allowed=paper_probation_allowed,
-            source_lineage_ok=source_lineage_ok,
-            exact_replay_parity_ok=exact_replay_parity_ok,
-            diagnostics=handoff_diagnostics,
-        )
-        net_pnl_per_day = _candidate_metric_decimal(
-            item,
-            scorecard_key="net_pnl_per_day",
-            full_window_key="net_per_day",
-        )
-        net_pnl = _candidate_metric_decimal(
-            item,
-            scorecard_key="net_pnl",
-            full_window_key="net_pnl",
-        )
-        capital_repair_notional_scale = _paper_probation_notional_scale_decimal(
-            item=item,
-            objective_veto_policy=objective_veto_policy,
-            hard_vetoes=hard_vetoes,
-        )
-        capital_repaired_net_pnl_per_day = (
-            net_pnl_per_day * capital_repair_notional_scale
-        )
-        capital_repaired_target_shortfall = max(
-            target_net_pnl_per_day - capital_repaired_net_pnl_per_day,
-            Decimal("0"),
-        )
-        target_notional_scale_after_capital_repair = (
-            _paper_probation_target_notional_scale(
-                capital_repaired_net_pnl_per_day=capital_repaired_net_pnl_per_day,
-                target_net_pnl_per_day=target_net_pnl_per_day,
-            )
-        )
-        target_shortfall = max(target_net_pnl_per_day - net_pnl_per_day, Decimal("0"))
-        repair_plan = _paper_probation_repair_plan(
-            item=item,
-            blockers=blockers,
-            hard_vetoes=hard_vetoes,
-            handoff_diagnostics=handoff_diagnostics,
-            net_pnl_per_day=net_pnl_per_day,
-            target_net_pnl_per_day=target_net_pnl_per_day,
-            capital_repair_notional_scale=capital_repair_notional_scale,
-            capital_repaired_net_pnl_per_day=capital_repaired_net_pnl_per_day,
-            target_notional_scale_after_capital_repair=(
-                target_notional_scale_after_capital_repair
-            ),
-        )
-        required_actions = _paper_probation_required_actions(hard_vetoes)
-        if target_notional_scale_after_capital_repair > Decimal("1"):
-            required_actions.append(
-                "validate_target_notional_scale_capacity_before_paper_orders"
-            )
-        entry = {
-            "candidate_id": str(item.get("candidate_id") or ""),
-            "strategy_name": str(item.get("strategy_name") or ""),
-            "family": str(item.get("family") or ""),
-            "stage": "paper_evidence_collection_only",
-            "paper_probation_allowed": paper_probation_allowed,
-            "evidence_collection_ok": bounded_sim_handoff["evidence_collection_ok"],
-            "promotion_allowed": False,
-            "final_promotion_allowed": False,
-            "final_promotion_authorized": False,
-            "probation_blockers": blockers,
-            "handoff_diagnostics": handoff_diagnostics,
-            "bounded_sim_handoff": bounded_sim_handoff,
-            "paper_probation_repair_plan": repair_plan,
-            "required_actions_before_or_during_probation": list(
-                dict.fromkeys(required_actions)
-            ),
-            "live_paper_evidence_requirements": repair_plan[
-                "live_paper_evidence_requirements"
-            ],
-            "safe_evidence_collection_path": repair_plan[
-                "safe_evidence_collection_path"
-            ],
-            "live_capital_authorized": False,
-            "recommended_notional_scale": _decimal_payload(
-                capital_repair_notional_scale
-            ),
-            "capital_repaired_net_pnl_per_day": _decimal_payload(
-                capital_repaired_net_pnl_per_day
-            ),
-            "capital_repaired_target_shortfall": _decimal_payload(
-                capital_repaired_target_shortfall
-            ),
-            "target_notional_scale_after_capital_repair": _decimal_payload(
-                target_notional_scale_after_capital_repair
-            ),
-            "target_scale_required": (
-                target_notional_scale_after_capital_repair > Decimal("1")
-            ),
-            "target_scale_authority": "planning_only_requires_bounded_paper_validation",
-            "exact_replay_ledger_artifact_refs": artifact_refs,
-            "exact_replay_ledger_artifact_row_count": (exact_replay_ledger_row_count),
-            "exact_replay_ledger_artifact_fill_count": (exact_replay_ledger_fill_count),
-            "target_net_pnl_per_day": _decimal_payload(target_net_pnl_per_day),
-            "target_shortfall": _decimal_payload(target_shortfall),
-            "target_progress_ratio": _decimal_payload(
-                _paper_probation_target_progress(
-                    net_pnl_per_day=net_pnl_per_day,
-                    target_net_pnl_per_day=target_net_pnl_per_day,
-                )
-            ),
-            "net_pnl_per_day": str(net_pnl_per_day),
-            "net_pnl": str(net_pnl),
-            "active_day_ratio": str(
-                _candidate_metric_value(
-                    item,
-                    scorecard_key="active_day_ratio",
-                    full_window_key="active_ratio",
-                )
-            ),
-            "positive_day_ratio": str(
-                _candidate_metric_value(
-                    item,
-                    scorecard_key="positive_day_ratio",
-                    full_window_key="positive_day_ratio",
-                )
-            ),
-            "max_drawdown": str(
-                _candidate_metric_value(
-                    item,
-                    scorecard_key="max_drawdown",
-                    full_window_key="max_drawdown",
-                )
-            ),
-            "worst_day_loss": str(
-                _candidate_metric_value(
-                    item,
-                    scorecard_key="worst_day_loss",
-                    full_window_key="worst_day_loss",
-                )
-            ),
-            "max_gross_exposure_pct_equity": str(
-                _candidate_metric_value(
-                    item,
-                    scorecard_key="max_gross_exposure_pct_equity",
-                    full_window_key="max_gross_exposure_pct_equity",
-                )
-            ),
-            "min_cash": str(
-                _candidate_metric_value(
-                    item,
-                    scorecard_key="min_cash",
-                    full_window_key="min_cash",
-                )
-            ),
-            "exact_replay_ledger_artifact_ref": str(
-                item.get("exact_replay_ledger_artifact_ref") or ""
-            ),
-            "replay_artifact_refs": artifact_refs,
-            "hard_vetoes": hard_vetoes,
-        }
-        shortlist_candidates.append(
-            (
-                paper_probation_allowed,
-                bool(bounded_sim_handoff["evidence_collection_ok"]),
-                capital_repaired_target_shortfall,
-                target_notional_scale_after_capital_repair,
-                net_pnl_per_day,
-                net_pnl,
-                entry["candidate_id"],
-                entry,
-            )
-        )
-
-    shortlist_candidates.sort(
-        key=lambda candidate: (
-            0 if candidate[0] else 1,
-            0 if candidate[1] else 1,
-            candidate[2],
-            candidate[3],
-            -candidate[4],
-            -candidate[5],
-            candidate[6],
-        )
-    )
-    return [candidate[7] for candidate in shortlist_candidates[: max(1, int(top_n))]]
-
-
-def _frontier_state_item(item: Mapping[str, Any]) -> dict[str, Any]:
-    screening = _mapping(item.get("screening"))
-    staged_search = _mapping(item.get("staged_search"))
-    ranking = _mapping(item.get("ranking"))
-    return {
-        "candidate_id": str(item.get("candidate_id") or ""),
-        "strategy_name": str(item.get("strategy_name") or ""),
-        "family": str(item.get("family") or ""),
-        "stage": str(staged_search.get("stage") or ""),
-        "screening_status": str(screening.get("status") or ""),
-        "train_screen_survivor_rank": staged_search.get("train_screen_survivor_rank"),
-        "exact_replay_selection_reason": staged_search.get(
-            "full_replay_selection_reason"
-        ),
-        "net_pnl_per_day": str(
-            _candidate_metric_value(
-                item,
-                scorecard_key="net_pnl_per_day",
-                full_window_key="net_per_day",
-            )
-        ),
-        "hard_vetoes": [
-            str(reason) for reason in cast(Sequence[Any], item.get("hard_vetoes") or ())
-        ],
-        "vetoed": bool(ranking.get("vetoed") or item.get("hard_vetoes")),
-    }
-
-
-def _build_frontier_workflow_states(
-    ranked_items: Sequence[Mapping[str, Any]],
-    *,
-    paper_probation_shortlist: Sequence[Mapping[str, Any]],
-) -> dict[str, Any]:
-    preview_qualified = [
-        _frontier_state_item(item)
-        for item in ranked_items
-        if str(_mapping(item.get("screening")).get("status") or "") == "passed"
-    ]
-    exact_replay_shortlist = [
-        _frontier_state_item(item)
-        for item in ranked_items
-        if bool(
-            _mapping(item.get("staged_search")).get(
-                "full_replay_selected_after_train_rank"
-            )
-        )
-        or str(_mapping(item.get("staged_search")).get("stage") or "") == "full_replay"
-    ]
-    exact_replay_qualified = [
-        {
-            **_frontier_state_item(item),
-            "exact_replay_ledger_artifact_refs": (
-                _candidate_exact_replay_ledger_artifact_refs(item)
-            ),
-        }
-        for item in ranked_items
-        if str(_mapping(item.get("staged_search")).get("stage") or "") == "full_replay"
-        and _candidate_runtime_ledger_count(
-            item,
-            "exact_replay_ledger_artifact_row_count",
-        )
-        > 0
-    ]
-    return {
-        "schema_version": "torghut.frontier-workflow-states.v1",
-        "purpose": (
-            "separate preview narrowing, capped exact replay, bounded SIM handoff, "
-            "and absent promotion authority"
-        ),
-        "preview_qualified": preview_qualified,
-        "exact_replay_shortlist": exact_replay_shortlist[
-            :_SAFE_EXACT_REPLAY_CANDIDATE_CAP
-        ],
-        "exact_replay_qualified": exact_replay_qualified,
-        "paper_probation_shortlisted": [
-            dict(item) for item in paper_probation_shortlist
-        ],
-        "paper_probation_repair_queue": [
-            {
-                "candidate_id": str(item.get("candidate_id") or ""),
-                "target_shortfall": str(item.get("target_shortfall") or "0"),
-                "target_progress_ratio": str(item.get("target_progress_ratio") or "0"),
-                "repair_plan": dict(_mapping(item.get("paper_probation_repair_plan"))),
-            }
-            for item in paper_probation_shortlist
-            if not bool(item.get("paper_probation_allowed"))
-            and bool(
-                _mapping(item.get("paper_probation_repair_plan")).get(
-                    "repairable_for_evidence_collection"
-                )
-            )
-        ],
-        "authority_proof": {
-            "status": "absent",
-            "promotion_allowed": False,
-            "final_promotion_allowed": False,
-            "proof_authority": False,
-            "authority_blockers": [
-                "source_backed_runtime_ledger_authority_required",
-                "live_paper_probation_runtime_evidence_required",
-                "runtime_ledger_to_broker_lineage_required",
-                "unchanged_final_promotion_gates_required",
-            ],
-        },
-    }
-
-
-def _build_frontier_payload(
-    *,
-    scored: list[dict[str, Any]],
-    family: str,
-    strategy_name: str,
-    family_template: Any,
-    dataset_snapshot_receipt: Any,
-    window: Any,
-    full_window_start: date,
-    full_window_end: date,
-    holdout_policy: ProfitabilityConstraintPolicy,
-    consistency_policy: FullWindowConsistencyPolicy,
-    objective_veto_policy: ObjectiveVetoPolicy,
-    top_n: int,
-    status: str,
-    pending_candidates: int,
-    replay_tape_validation: Mapping[str, Any] | None = None,
-    staged_search: Mapping[str, Any] | None = None,
-) -> dict[str, Any]:
-    ranked_items = _rank_scored_candidates(scored)
-    economic_shortlist_items = _build_economic_shortlist(
-        ranked_items,
-        top_n=max(1, int(top_n)),
-    )
-    paper_probation_shortlist_items = _build_paper_probation_shortlist(
-        ranked_items,
-        top_n=max(3, int(top_n)),
-        objective_veto_policy=objective_veto_policy,
-        target_net_pnl_per_day=consistency_policy.target_net_per_day,
-    )
-    return {
-        "schema_version": _SWEEP_SCHEMA_VERSION,
-        "status": status,
-        "family": family,
-        "strategy_name": strategy_name,
-        "family_template": family_template.to_payload(),
-        "dataset_snapshot_receipt": dataset_snapshot_receipt.to_payload(),
-        "window": {
-            "train_days": [item.isoformat() for item in window.train_days],
-            "holdout_days": [item.isoformat() for item in window.holdout_days],
-            "second_oos_days": [item.isoformat() for item in window.second_oos_days],
-            "full_window_start_date": full_window_start.isoformat(),
-            "full_window_end_date": full_window_end.isoformat(),
-        },
-        "constraints": {
-            "holdout": {
-                "holdout_target_net_per_day": str(
-                    holdout_policy.holdout_target_net_per_day
-                ),
-                "min_active_holdout_days": holdout_policy.min_active_holdout_days,
-                "max_worst_holdout_day_loss": str(
-                    holdout_policy.max_worst_holdout_day_loss
-                ),
-                "min_profit_factor": str(holdout_policy.min_profit_factor),
-                "require_training_decisions": holdout_policy.require_training_decisions,
-                "require_holdout_decisions": holdout_policy.require_holdout_decisions,
-            },
-            "second_oos": {
-                "enabled": bool(window.second_oos_days),
-                "window_id": _SECOND_OOS_WINDOW_ID,
-                "min_independent_window_count": 2 if window.second_oos_days else 1,
-                "target_net_per_day": str(consistency_policy.target_net_per_day),
-            },
-            "consistency": consistency_policy.to_payload(),
-            "hard_vetoes": objective_veto_policy.to_payload(),
-        },
-        "ranking": {
-            "method": "pareto_frontier_v2",
-            "stale_override_used": dataset_snapshot_receipt.stale_override_used,
-        },
-        "replay_tape": dict(replay_tape_validation or {}),
-        "progress": {
-            "evaluated_candidates": len(scored),
-            "pending_candidates": pending_candidates,
-            "staged_search": dict(staged_search or {}),
-        },
-        "candidate_count": len(scored),
-        "economic_shortlist": {
-            "schema_version": "torghut.frontier-economic-shortlist.v1",
-            "ranking_basis": "full_window_net_pnl_per_day_desc",
-            "items": economic_shortlist_items,
-        },
-        "paper_probation_shortlist": {
-            "schema_version": "torghut.frontier-paper-probation-shortlist.v1",
-            "purpose": (
-                "surface close positive candidates for controlled paper evidence "
-                "collection without authorizing final promotion"
-            ),
-            "ranking_basis": "paper_readiness_then_target_shortfall_then_net_pnl",
-            "items": paper_probation_shortlist_items,
-        },
-        "workflow_states": _build_frontier_workflow_states(
-            ranked_items,
-            paper_probation_shortlist=paper_probation_shortlist_items,
-        ),
-        "top": ranked_items[: max(1, int(top_n))],
-    }
 
 
 def _selected_normalization_regime(
